Reorganized types
